# artifacts/api-server/src/routes/reports.ts
import { Router, type IRouter } from "express";
import { db, customersTable, saleOrdersTable, saleOrderItemsTable, paymentsTable, productsTable, expensesTable } from "@workspace/db";
import { saleReturnsTable, saleReturnItemsTable, customerLoansTable } from "@workspace/db/schema";
import { eq, sql, and, gte, lte } from "drizzle-orm";
import { requireAuth } from "../middlewares/auth";
import { computeCustomerBalance } from "../lib/customerBalance";
import { GetDailyCollectionReportQueryParams, GetMonthlySalesReportQueryParams } from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/reports/aging", requireAuth, async (_req, res): Promise<void> => {
  const customers = await db.select().from(customersTable).orderBy(customersTable.name);
  const now = new Date();

  const result = await Promise.all(customers.map(async (c) => {
    const orders = await db.select().from(saleOrdersTable).where(and(eq(saleOrdersTable.customerId, c.id), eq(saleOrdersTable.status, "posted")));
    const pmts = await db.select({ total: sql<number>`coalesce(sum(${paymentsTable.amount}),0)` }).from(paymentsTable).where(and(eq(paymentsTable.customerId, c.id), eq(paymentsTable.status, "posted")));
    // Sale returns net against the specific order they were made against (each return
    // is tied to one saleOrderId) — a return reduces that order's effective amount, a
    // cash refund on top of it adds back, per the balance-formula worked example in
    // customers.ts's computeCustomerBalance.
    const returns = await db.select().from(saleReturnsTable).where(and(eq(saleReturnsTable.customerId, c.id), eq(saleReturnsTable.status, "posted")));
    const returnsByOrder = new Map<number, { total: number; refunded: number }>();
    for (const r of returns) {
      const cur = returnsByOrder.get(r.saleOrderId) ?? { total: 0, refunded: 0 };
      cur.total += parseFloat(r.totalAmount);
      cur.refunded += parseFloat(r.refundPaid);
      returnsByOrder.set(r.saleOrderId, cur);
    }

    // Loans increase what the customer owes, same direction as a sale — see the
    // balance-formula rationale in customers.ts's computeCustomerBalance. Aged by the
    // loan's own date, same as a sale order, so they land in the correct day bucket too.
    const loans = await db.select().from(customerLoansTable).where(and(eq(customerLoansTable.customerId, c.id), eq(customerLoansTable.status, "posted")));

    const openingBal = parseFloat(c.openingBalance ?? "0");
    const totalPaid = parseFloat(String(pmts[0]?.total ?? 0));
    let remaining = totalPaid;

    // Discount is aged by the order's own date, same as the sale itself, since it's
    // agreed at sale time — see saleOrders.ts's discountAmount column comment.
    const items = [
      ...orders.map((o) => {
        const ret = returnsByOrder.get(o.id);
        const discount = parseFloat(o.discountAmount ?? "0");
        return { date: o.date, amt: parseFloat(o.totalAmount) - (ret?.total ?? 0) + (ret?.refunded ?? 0) - discount };
      }),
      ...loans.map((l) => ({ date: l.date, amt: parseFloat(l.amount) })),
    ];

    let d0to30 = 0, d31to60 = 0, d61to90 = 0, dOver90 = 0;
    for (const item of items.sort((a, b) => a.date.localeCompare(b.date))) {
      const days = Math.floor((now.getTime() - new Date(item.date).getTime()) / (1000 * 60 * 60 * 24));
      const unpaid = Math.max(0, item.amt - remaining);
      remaining = Math.max(0, remaining - item.amt);
      if (days <= 30) d0to30 += unpaid;
      else if (days <= 60) d31to60 += unpaid;
      else if (days <= 90) d61to90 += unpaid;
      else dOver90 += unpaid;
    }

    // Sourced from the same shared formula the Customers page uses, so this always
    // agrees with it exactly (previously hand-duplicated here and it drifted — see the
    // customer-loans balance bug this replaced).
    const balance = await computeCustomerBalance(c.id, openingBal);
    if (balance <= 0) return null;

    return {
      customerId: c.id, customerName: c.name, area: c.area ?? null, contact: c.contact ?? null,
      balance, days0to30: d0to30, days31to60: d31to60, days61to90: d61to90, daysOver90: dOver90,
    };
  }));

  res.json(result.filter(Boolean));
});

router.get("/reports/daily-collection", requireAuth, async (req, res): Promise<void> => {
  const query = GetDailyCollectionReportQueryParams.safeParse(req.query);
  const date = (query.success && query.data.date) ? query.data.date : new Date().toISOString().split("T")[0];

  const pmts = await db.select().from(paymentsTable).where(and(eq(paymentsTable.date, String(date)), eq(paymentsTable.status, "posted"))).orderBy(paymentsTable.createdAt);
  const result = await Promise.all(pmts.map(async (p) => {
    const [c] = await db.select().from(customersTable).where(eq(customersTable.id, p.customerId));
    return { id: p.id, customerId: p.customerId, customerName: c?.name ?? "", date: p.date, type: p.type, amount: parseFloat(p.amount), bankAccount: p.bankAccount ?? null, chequeNo: p.chequeNo ?? null, notes: p.notes ?? null, createdAt: p.createdAt };
  }));

  const totalCash = result.filter(p => p.type === "cash").reduce((s, p) => s + p.amount, 0);
  const totalBank = result.filter(p => p.type === "bank").reduce((s, p) => s + p.amount, 0);
  res.json({ date, totalCash, totalBank, total: totalCash + totalBank, payments: result });
});

router.get("/reports/monthly-sales", requireAuth, async (req, res): Promise<void> => {
  const query = GetMonthlySalesReportQueryParams.safeParse(req.query);
  const now = new Date();
  const year = (query.success && query.data.year) ? query.data.year : now.getFullYear();
  const month = (query.success && query.data.month) ? query.data.month : now.getMonth() + 1;

  const from = `${year}-${String(month).padStart(2, "0")}-01`;
  const toDate = new Date(year, month, 0);
  const to = `${year}-${String(month).padStart(2, "0")}-${String(toDate.getDate()).padStart(2, "0")}`;

  const orders = await db.select().from(saleOrdersTable).where(and(gte(saleOrdersTable.date, from), lte(saleOrdersTable.date, to), eq(saleOrdersTable.status, "posted")));
  const products = await db.select().from(productsTable);

  const byProduct: Record<number, { productId: number; productName: string; totalQty: number; totalAmount: number }> = {};
  let totalQty = 0, totalAmount = 0;

  for (const o of orders) {
    const items = await db.select().from(saleOrderItemsTable).where(eq(saleOrderItemsTable.saleOrderId, o.id));
    for (const item of items) {
      const product = products.find(p => p.id === item.productId);
      if (!byProduct[item.productId]) byProduct[item.productId] = { productId: item.productId, productName: product?.name ?? "", totalQty: 0, totalAmount: 0 };
      byProduct[item.productId].totalQty += parseFloat(item.qty);
      byProduct[item.productId].totalAmount += parseFloat(item.amount);
      totalQty += parseFloat(item.qty);
    }
    totalAmount += parseFloat(o.totalAmount);
  }

  res.json({ year, month, totalAmount, totalQty, byProduct: Object.values(byProduct) });
});

router.get("/reports/daily-profit", requireAuth, async (req, res): Promise<void> => {
  const now = new Date();
  const firstOfMonth = `${now.getFullYear()}-${String(now.getMonth() + 1).padStart(2, "0")}-01`;
  const todayStr = now.toISOString().slice(0, 10);
  const from = (req.query.from as string) || firstOfMonth;
  const to = (req.query.to as string) || todayStr;
  const categoryFilter = req.query.category ? String(req.query.category) : null;

  // Fetch all sale orders in range with their items — only "live" (posted) orders, so
  // this reflects corrections, not mistakes (see the correction workflow).
  const orders = await db
    .select()
    .from(saleOrdersTable)
    .where(and(gte(saleOrdersTable.date, from), lte(saleOrdersTable.date, to), eq(saleOrdersTable.status, "posted")));

  const allProducts = await db.select().from(productsTable);
  // Build a set of product IDs matching the category filter
  const filteredProductIds = categoryFilter
    ? new Set(allProducts.filter(p => p.category === categoryFilter).map(p => p.id))
    : null;
  const productMap = new Map(allProducts.map(p => [p.id, p]));

  // Fetch all expenses in range
  const expenses = await db
    .select()
    .from(expensesTable)
    .where(and(gte(expensesTable.date, from), lte(expensesTable.date, to)));

  // Build day map
  const dayMap = new Map<string, {
    revenue: number; cogs: number; expenses: number; orders: number; qty: number;
  }>();

  // Build product map for breakdown
  const byProduct = new Map<number, {
    productId: number; productName: string; qty: number; revenue: number; cogs: number;
  }>();

  for (const order of orders) {
    const dateStr = typeof order.date === "string" ? order.date : (order.date as Date).toISOString().slice(0, 10);
    if (!dayMap.has(dateStr)) {
      dayMap.set(dateStr, { revenue: 0, cogs: 0, expenses: 0, orders: 0, qty: 0 });
    }
    const day = dayMap.get(dateStr)!;

    const items = await db
      .select()
      .from(saleOrderItemsTable)
      .where(eq(saleOrderItemsTable.saleOrderId, order.id));

    let orderHasMatchingItems = false;
    for (const item of items) {
      // Skip items not in the filtered category (when a filter is active)
      if (filteredProductIds && !filteredProductIds.has(item.productId)) continue;

      orderHasMatchingItems = true;
      const qty = parseFloat(item.qty);
      const saleAmount = parseFloat(item.amount);
      const product = productMap.get(item.productId);
      // Prefer the frozen per-sale snapshot; fall back to the product's live cost price
      // only for legacy rows created before cost price was captured at sale time.
      const costPriceRaw = item.costPrice ?? product?.costPrice ?? null;
      const costPrice = costPriceRaw ? parseFloat(costPriceRaw) : 0;
      const itemCogs = qty * costPrice;

      day.qty += qty;
      day.cogs += itemCogs;
      day.revenue += saleAmount;

      // Product breakdown
      if (!byProduct.has(item.productId)) {
        byProduct.set(item.productId, {
          productId: item.productId,
          productName: product?.name ?? "",
          qty: 0, revenue: 0, cogs: 0,
        });
      }
      const prod = byProduct.get(item.productId)!;
      prod.qty += qty;
      prod.revenue += saleAmount;
      prod.cogs += itemCogs;
    }

    // When no category filter, add full order revenue (avoids rounding discrepancies),
    // net of any sale-time discount — pure revenue given up, not tied to one product.
    // When a filter is active, revenue is already accumulated per item above (and the
    // discount isn't attributed to any single category, so it's left out there).
    if (!filteredProductIds) {
      day.revenue += parseFloat(order.totalAmount) - parseFloat(order.discountAmount ?? "0");
      // Undo the per-item revenue we added (to avoid double-counting)
      for (const item of items) {
        day.revenue -= parseFloat(item.amount);
      }
    }
    if (orderHasMatchingItems || !filteredProductIds) {
      day.orders += 1;
    }
  }

  // Sale returns reverse revenue/COGS on the return's own date (not retroactively on
  // the original sale's date) — mirrors the loop above, subtracting instead of adding,
  // using each return item's frozen costPrice snapshot (copied from the original sale
  // item at return time, not re-fetched from the product, so this matches exactly what
  // was booked as profit on the original sale).
  const returns = await db
    .select()
    .from(saleReturnsTable)
    .where(and(gte(saleReturnsTable.date, from), lte(saleReturnsTable.date, to), eq(saleReturnsTable.status, "posted")));

  for (const ret of returns) {
    const dateStr = typeof ret.date === "string" ? ret.date : (ret.date as Date).toISOString().slice(0, 10);
    if (!dayMap.has(dateStr)) {
      dayMap.set(dateStr, { revenue: 0, cogs: 0, expenses: 0, orders: 0, qty: 0 });
    }
    const day = dayMap.get(dateStr)!;

    const items = await db.select().from(saleReturnItemsTable).where(eq(saleReturnItemsTable.saleReturnId, ret.id));
    for (const item of items) {
      if (filteredProductIds && !filteredProductIds.has(item.productId)) continue;

      const qty = parseFloat(item.qty);
      const returnAmount = parseFloat(item.amount);
      const costPrice = item.costPrice ? parseFloat(item.costPrice) : 0;
      const itemCogs = qty * costPrice;

      day.qty -= qty;
      day.cogs -= itemCogs;
      day.revenue -= returnAmount;

      const product = productMap.get(item.productId);
      if (!byProduct.has(item.productId)) {
        byProduct.set(item.productId, {
          productId: item.productId,
          productName: product?.name ?? "",
          qty: 0, revenue: 0, cogs: 0,
        });
      }
      const prod = byProduct.get(item.productId)!;
      prod.qty -= qty;
      prod.revenue -= returnAmount;
      prod.cogs -= itemCogs;
    }
  }

  // Distribute expenses into day map
  for (const exp of expenses) {
    const dateStr = typeof exp.date === "string" ? exp.date : (exp.date as Date).toISOString().slice(0, 10);
    if (!dayMap.has(dateStr)) {
      dayMap.set(dateStr, { revenue: 0, cogs: 0, expenses: 0, orders: 0, qty: 0 });
    }
    dayMap.get(dateStr)!.expenses += parseFloat(exp.amount);
  }

  // Sort days
  const days = Array.from(dayMap.entries())
    .sort(([a], [b]) => a.localeCompare(b))
    .map(([date, d]) => {
      const grossProfit = d.revenue - d.cogs;
      const netProfit = grossProfit - d.expenses;
      return {
        date,
        revenue: Math.round(d.revenue * 100) / 100,
        cogs: Math.round(d.cogs * 100) / 100,
        grossProfit: Math.round(grossProfit * 100) / 100,
        grossMargin: d.revenue > 0 ? Math.round((grossProfit / d.revenue) * 10000) / 100 : 0,
        expenses: Math.round(d.expenses * 100) / 100,
        netProfit: Math.round(netProfit * 100) / 100,
        netMargin: d.revenue > 0 ? Math.round((netProfit / d.revenue) * 10000) / 100 : 0,
        orders: d.orders,
        qty: Math.round(d.qty * 100) / 100,
      };
    });

  // Totals
  const totRevenue = days.reduce((s, d) => s + d.revenue, 0);
  const totCogs = days.reduce((s, d) => s + d.cogs, 0);
  const totExpenses = days.reduce((s, d) => s + d.expenses, 0);
  const totGross = totRevenue - totCogs;
  const totNet = totGross - totExpenses;
  const totOrders = days.reduce((s, d) => s + d.orders, 0);
  const totQty = days.reduce((s, d) => s + d.qty, 0);

  const summary = {
    revenue: Math.round(totRevenue * 100) / 100,
    cogs: Math.round(totCogs * 100) / 100,
    grossProfit: Math.round(totGross * 100) / 100,
    grossMargin: totRevenue > 0 ? Math.round((totGross / totRevenue) * 10000) / 100 : 0,
    expenses: Math.round(totExpenses * 100) / 100,
    netProfit: Math.round(totNet * 100) / 100,
    netMargin: totRevenue > 0 ? Math.round((totNet / totRevenue) * 10000) / 100 : 0,
    orders: totOrders,
    qty: Math.round(totQty * 100) / 100,
  };

  const byProductArr = Array.from(byProduct.values()).map(p => ({
    ...p,
    qty: Math.round(p.qty * 100) / 100,
    revenue: Math.round(p.revenue * 100) / 100,
    cogs: Math.round(p.cogs * 100) / 100,
    profit: Math.round((p.revenue - p.cogs) * 100) / 100,
    margin: p.revenue > 0 ? Math.round(((p.revenue - p.cogs) / p.revenue) * 10000) / 100 : 0,
  })).sort((a, b) => b.profit - a.profit);

  res.json({ from, to, summary, days, byProduct: byProductArr });
});

router.get("/reports/outstanding", requireAuth, async (_req, res): Promise<void> => {
  const customers = await db.select().from(customersTable).orderBy(customersTable.name);
  const result = await Promise.all(customers.map(async (c) => {
    // Sourced from the same shared formula the Customers page uses — see aging above
    // for why this isn't hand-duplicated per report anymore.
    const balance = await computeCustomerBalance(c.id, c.openingBalance ?? "0");

    const lastSale = await db.select({ date: saleOrdersTable.date }).from(saleOrdersTable).where(and(eq(saleOrdersTable.customerId, c.id), eq(saleOrdersTable.status, "posted"))).orderBy(sql`${saleOrdersTable.date} desc`).limit(1);
    const lastPmt = await db.select({ date: paymentsTable.date }).from(paymentsTable).where(and(eq(paymentsTable.customerId, c.id), eq(paymentsTable.status, "posted"))).orderBy(sql`${paymentsTable.date} desc`).limit(1);
    const creditLimit = c.creditLimit ? parseFloat(c.creditLimit) : null;
    const isOverLimit = creditLimit != null && balance > creditLimit;

    return {
      customerId: c.id, customerName: c.name, area: c.area ?? null, contact: c.contact ?? null,
      creditLimit, balance, lastSaleDate: lastSale[0]?.date ?? null, lastPaymentDate: lastPmt[0]?.date ?? null, isOverLimit,
    };
  }));
  res.json(result.filter(r => r.balance > 0));
});

export default router;