# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AL-RAHMAN TRADERS Ledger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/daily-collection:
    get:
      operationId: getDailyCollectionReport
      tags: [reports]
      summary: Daily collection summary
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyCollectionReport"
  /reports/monthly-sales:
    get:
      operationId: getMonthlySalesReport
      tags: [reports]
      summary: Monthly sales by product
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlySalesReport"
  /reports/daily-profit:
    get:
      operationId: getDailyProfitReport
      tags: [reports]
      summary: Daily profit report — revenue, COGS, gross profit, expenses, net profit
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter results to a specific product category
      responses:
        "200":
          description: Daily profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyProfitReport"
...
components:
  schemas:
...
    Payment:
      type: object
      required: [id, customerId, customerName, date, type, amount, createdAt, status]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
          enum: [cash, bank]
        amount:
          type: number
        bankAccount:
          type: ["string", "null"]
        chequeNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [posted, reversed, reversal]
        reversesId:
          type: ["integer", "null"]
        correctsId:
          type: ["integer", "null"]
...
    DailyCollectionReport:
      type: object
      required: [date, totalCash, totalBank, total, payments]
      properties:
        date:
          type: string
          format: date
        totalCash:
          type: number
        totalBank:
          type: number
        total:
          type: number
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    MonthlySalesProductRow:
      type: object
      required: [productId, productName, totalQty, totalAmount]
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQty:
          type: number
        totalAmount:
          type: number
    MonthlySalesReport:
      type: object
      required: [year, month, totalAmount, totalQty, byProduct]
      properties:
        year:
          type: integer
        month:
          type: integer
        totalAmount:
          type: number
        totalQty:
          type: number
        byProduct:
          type: array
          items:
            $ref: "#/components/schemas/MonthlySalesProductRow"
...
    DailyProfitDay:
      type: object
      required: [date, revenue, cogs, grossProfit, grossMargin, expenses, netProfit, netMargin, orders, qty]
      properties:
        date:
          type: string
          format: date
        revenue:
          type: number
        cogs:
          type: number
        grossProfit:
          type: number
        grossMargin:
          type: number
        expenses:
          type: number
        netProfit:
          type: number
        netMargin:
          type: number
        orders:
          type: integer
        qty:
          type: number
    DailyProfitProductRow:
      type: object
      required: [productId, productName, qty, revenue, cogs, profit, margin]
      properties:
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: number
        revenue:
          type: number
        cogs:
          type: number
        profit:
          type: number
        margin:
          type: number
    DailyProfitSummary:
      type: object
      required: [revenue, cogs, grossProfit, grossMargin, expenses, netProfit, netMargin, orders, qty]
      properties:
        revenue:
          type: number
        cogs:
          type: number
        grossProfit:
          type: number
        grossMargin:
          type: number
        expenses:
          type: number
        netProfit:
          type: number
        netMargin:
          type: number
        orders:
          type: integer
        qty:
          type: number
    DailyProfitReport:
      type: object
      required: [from, to, summary, days, byProduct]
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        summary:
          $ref: "#/components/schemas/DailyProfitSummary"
        days:
          type: array
          items:
            $ref: "#/components/schemas/DailyProfitDay"
        byProduct:
          type: array
          items:
            $ref: "#/components/schemas/DailyProfitProductRow"

    # ── SUPPLIER / PURCHASE SCHEMAS ───────────────────────────────────────
...